// AIMS-BE/src/data-access-layer/daos/products/interfaces/book-model.dto.ts
import { COVER_TYPE } from '../../../../configs/enums'
import { IProduct } from './product.interface'

export interface IBook extends IProduct {
    id: number | ObjectId | string
    author: string
    coverType: COVER_TYPE
    publisher: string
    publicationDate: Date
    pages: number
    language: string
    bookCategory: string
}